Flesh Out tails_introduction.md
This commit creates a skeleton doc that ppl can edit. Bare intro to what Tails is, and why you'd want to use it. The main section of the intro just lists the most important features of Tails:

- runs off external storage
- securely wipes itself upon shutdown
- crypto tools come as standard
- all connections go through Tor

Leads with quote from LP on the importance of Tails from the FPF blogpost calling for donations.